restructuring event modal
